update mgt-react to use wc-react 0.4, react sample is now ts
