+++ packages/mgt-react/package.json
@@ -33,9 +33,9 @@
   "dependencies": {
     "@microsoft/mgt-components": "*",
     "@microsoft/mgt-element": "*",
-    "@microsoft/microsoft-graph-types": "^1.27.0",
+    "@microsoft/microsoft-graph-types": "^1.28.0",
     "@microsoft/microsoft-graph-types-beta": "github:microsoftgraph/msgraph-typescript-typings#beta",
-    "wc-react": "^0.3.1"
+    "wc-react": "^0.4.0"
   },
   "publishConfig": {
     "access": "public"
