+++ samples/react-app/package.json
@@ -3,8 +3,9 @@
   "version": "1.3.0",
   "private": true,
   "dependencies": {
-    "@microsoft/mgt": "*",
+    "@microsoft/mgt-element": "*",
     "@microsoft/mgt-react": "*",
+    "@microsoft/mgt-msal-provider": "*",
     "@webcomponents/webcomponentsjs": "2.2.10",
     "react": "^16.8.5",
     "react-app-polyfill": "1.0.1",
